refactor: Add EmptyProjectList component and update Projects template
